# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PawSmart API — AI Pet Care + E-Commerce
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  name: string;
  brand: string;
  /** food | treats | accessories | grooming | toys */
  category: string;
  /** dog | cat | small_pet */
  petType: string;
  price: number;
  stock: number;
  description: string;
  image: string;
  /** e.g. 1kg, 5kg, 250g */
  weight: string;
  expiryDate: string;
  popular: boolean;
  breed?: string | null;
  /** puppy | adult | senior | kitten | all */
  ageGroup?: string | null;
  healthCondition?: string | null;
  createdAt: string;
}

export interface ProductInput {
  name: string;
  brand: string;
  category: string;
  petType: string;
  price: number;
  stock: number;
  description: string;
  image: string;
  weight: string;
  expiryDate: string;
  popular?: boolean;
  breed?: string | null;
  ageGroup?: string | null;
  healthCondition?: string | null;
}

export interface CategorySummary {
  category: string;
  count: number;
}

export interface BrandSummary {
  brand: string;
  count: number;
}

export interface Pet {
  id: number;
  name: string;
  /** dog | cat */
  type: string;
  breed: string;
  ageYears: number;
  weightKg: number;
  healthNotes: string;
  createdAt: string;
}

export interface PetInput {
  name: string;
  type: string;
  breed: string;
  ageYears: number;
  weightKg: number;
  healthNotes?: string;
}

export interface CartItem {
  productId: number;
  quantity: number;
  product: Product;
}

export interface Cart {
  items: CartItem[];
  subtotal: number;
  shipping: number;
  total: number;
  count: number;
}

export interface AddCartItemInput {
  productId: number;
  /** @minimum 1 */
  quantity: number;
}

export interface UpdateCartItemInput {
  /** @minimum 0 */
  quantity: number;
}

export type BreedDetectInputPetType =
  (typeof BreedDetectInputPetType)[keyof typeof BreedDetectInputPetType];

export const BreedDetectInputPetType = {
  dog: "dog",
  cat: "cat",
} as const;

export interface BreedDetectInput {
  imageUrl: string;
  petType: BreedDetectInputPetType;
}

export type UpdateOrderStatusInputStatus =
  (typeof UpdateOrderStatusInputStatus)[keyof typeof UpdateOrderStatusInputStatus];

export const UpdateOrderStatusInputStatus = {
  pending: "pending",
  shipped: "shipped",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface UpdateOrderStatusInput {
  status: UpdateOrderStatusInputStatus;
}

export interface OrderItem {
  productId: number;
  name: string;
  brand: string;
  image: string;
  price: number;
  quantity: number;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  shipped: "shipped",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface Order {
  id: number;
  items: OrderItem[];
  subtotal: number;
  shipping: number;
  total: number;
  status: OrderStatus;
  customerName: string;
  address: string;
  phone: string;
  createdAt: string;
}

export interface CreateOrderInput {
  customerName: string;
  address: string;
  phone: string;
  notes?: string;
}

export interface AiRecommendInput {
  petType: string;
  breed?: string;
  ageYears?: number;
  weightKg?: number;
  healthCondition?: string;
}

export interface AiRecommendResult {
  summary: string;
  products: Product[];
  careTips: string[];
}

export interface SymptomCheckInput {
  petType: string;
  breed?: string;
  symptoms: string;
}

export type PossibleCauseSeverity =
  (typeof PossibleCauseSeverity)[keyof typeof PossibleCauseSeverity];

export const PossibleCauseSeverity = {
  mild: "mild",
  moderate: "moderate",
  severe: "severe",
} as const;

export interface PossibleCause {
  name: string;
  severity: PossibleCauseSeverity;
  description: string;
}

export interface SymptomCheckResult {
  disclaimer: string;
  possibleCauses: PossibleCause[];
  suggestedActions: string[];
  urgent: boolean;
}

export interface BreedDetectResult {
  breed: string;
  confidence: number;
  characteristics: string[];
  recommendedProducts: Product[];
  careTips: string[];
  disclaimer: string;
}

export type AdminStatsOrdersByStatusItem = {
  status: string;
  count: number;
};

export type AdminStatsRevenueByDayItem = {
  date: string;
  revenue: number;
};

export type AdminStatsTopProductsItem = {
  productId: number;
  name: string;
  totalSold: number;
  revenue: number;
};

export interface AdminStats {
  totalRevenue: number;
  totalOrders: number;
  totalProducts: number;
  totalUsers: number;
  lowStockCount: number;
  ordersByStatus: AdminStatsOrdersByStatusItem[];
  revenueByDay: AdminStatsRevenueByDayItem[];
  topProducts: AdminStatsTopProductsItem[];
}

export type ActivityItemType =
  (typeof ActivityItemType)[keyof typeof ActivityItemType];

export const ActivityItemType = {
  order: "order",
  user: "user",
  product: "product",
} as const;

export interface ActivityItem {
  type: ActivityItemType;
  title: string;
  subtitle: string;
  createdAt: string;
}

export type AdminUserRole = (typeof AdminUserRole)[keyof typeof AdminUserRole];

export const AdminUserRole = {
  user: "user",
  admin: "admin",
  vendor: "vendor",
} as const;

export interface AdminUser {
  id: number;
  name: string;
  email: string;
  role: AdminUserRole;
  blocked: boolean;
  createdAt: string;
  orderCount: number;
}

export type ListProductsParams = {
  category?: string;
  petType?: string;
  brand?: string;
  search?: string;
  sort?: ListProductsSort;
};

export type ListProductsSort =
  (typeof ListProductsSort)[keyof typeof ListProductsSort];

export const ListProductsSort = {
  popular: "popular",
  price_asc: "price_asc",
  price_desc: "price_desc",
  newest: "newest",
} as const;