# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PawSmart API — AI Pet Care + E-Commerce
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ai/recommend:
    post:
      operationId: aiRecommend
      tags: [ai]
      summary: Get product recommendations for a pet profile
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AiRecommendInput" }
      responses:
        "200":
          description: Recommendations
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AiRecommendResult" }
  /ai/symptom-check:
    post:
      operationId: aiSymptomCheck
      tags: [ai]
      summary: Symptom checker (with safety disclaimer)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SymptomCheckInput" }
      responses:
        "200":
          description: Symptom analysis
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SymptomCheckResult" }
  /ai/breed-detect:
    post:
      operationId: aiBreedDetect
      tags: [ai]
      summary: Pet image breed detector (mock identification)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BreedDetectInput" }
      responses:
        "200":
          description: Breed detection result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BreedDetectResult" }
...
  /admin/products:
    get:
      operationId: adminListProducts
      tags: [admin]
      responses:
        "200":
          description: All products
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Product" }
    post:
      operationId: adminCreateProduct
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProductInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
  /admin/products/{id}:
    put:
      operationId: adminUpdateProduct
      tags: [admin]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProductInput" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
    delete:
      operationId: adminDeleteProduct
      tags: [admin]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204": { description: Deleted }
...
  /admin/orders/{id}/status:
    put:
      operationId: adminUpdateOrderStatus
      tags: [admin]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateOrderStatusInput" }
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
...
components:
  schemas:
...
    Product:
      type: object
      required: [id, name, brand, category, petType, price, stock, description, image, weight, expiryDate, popular, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        brand: { type: string }
        category: { type: string, description: "food | treats | accessories | grooming | toys" }
        petType: { type: string, description: "dog | cat | small_pet" }
        price: { type: number }
        stock: { type: integer }
        description: { type: string }
        image: { type: string }
        weight: { type: string, description: "e.g. 1kg, 5kg, 250g" }
        expiryDate: { type: string }
        popular: { type: boolean }
        breed:
          type: string
          nullable: true
        ageGroup:
          type: string
          nullable: true
          description: "puppy | adult | senior | kitten | all"
        healthCondition:
          type: string
          nullable: true
        createdAt: { type: string }
    ProductInput:
      type: object
      required: [name, brand, category, petType, price, stock, description, image, weight, expiryDate]
      properties:
        name: { type: string }
        brand: { type: string }
        category: { type: string }
        petType: { type: string }
        price: { type: number }
        stock: { type: integer }
        description: { type: string }
        image: { type: string }
        weight: { type: string }
        expiryDate: { type: string }
        popular: { type: boolean }
        breed:
          type: string
          nullable: true
        ageGroup:
          type: string
          nullable: true
        healthCondition:
          type: string
          nullable: true
...
    BreedDetectInput:
      type: object
      required: [imageUrl, petType]
      properties:
        imageUrl: { type: string }
        petType: { type: string, enum: [dog, cat] }
    UpdateOrderStatusInput:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [pending, shipped, delivered, cancelled] }
    OrderItem:
      type: object
      required: [productId, name, brand, image, price, quantity]
      properties:
        productId: { type: integer }
        name: { type: string }
        brand: { type: string }
        image: { type: string }
        price: { type: number }
        quantity: { type: integer }
    Order:
      type: object
      required: [id, items, subtotal, shipping, total, status, customerName, address, phone, createdAt]
      properties:
        id: { type: integer }
        items:
          type: array
          items: { $ref: "#/components/schemas/OrderItem" }
        subtotal: { type: number }
        shipping: { type: number }
        total: { type: number }
        status: { type: string, enum: [pending, shipped, delivered, cancelled] }
        customerName: { type: string }
        address: { type: string }
        phone: { type: string }
        createdAt: { type: string }
...
    AiRecommendInput:
      type: object
      required: [petType]
      properties:
        petType: { type: string }
        breed: { type: string }
        ageYears: { type: number }
        weightKg: { type: number }
        healthCondition: { type: string }
    AiRecommendResult:
      type: object
      required: [summary, products, careTips]
      properties:
        summary: { type: string }
        products:
          type: array
          items: { $ref: "#/components/schemas/Product" }
        careTips:
          type: array
          items: { type: string }
    SymptomCheckInput:
      type: object
      required: [petType, symptoms]
      properties:
        petType: { type: string }
        breed: { type: string }
        symptoms: { type: string }
    PossibleCause:
      type: object
      required: [name, severity, description]
      properties:
        name: { type: string }
        severity: { type: string, enum: [mild, moderate, severe] }
        description: { type: string }
    SymptomCheckResult:
      type: object
      required: [disclaimer, possibleCauses, suggestedActions, urgent]
      properties:
        disclaimer: { type: string }
        possibleCauses:
          type: array
          items: { $ref: "#/components/schemas/PossibleCause" }
        suggestedActions:
          type: array
          items: { type: string }
        urgent: { type: boolean }
    BreedDetectResult:
      type: object
      required: [breed, confidence, characteristics, recommendedProducts, careTips, disclaimer]
      properties:
        breed: { type: string }
        confidence: { type: number }
        characteristics:
          type: array
          items: { type: string }
        recommendedProducts:
          type: array
          items: { $ref: "#/components/schemas/Product" }
        careTips:
          type: array
          items: { type: string }
        disclaimer: { type: string }
...