# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PawSmart API — AI Pet Care + E-Commerce
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from "./orderItem";
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: number;
  items: OrderItem[];
  subtotal: number;
  shipping: number;
  total: number;
  status: OrderStatus;
  customerName: string;
  address: string;
  phone: string;
  createdAt: string;
}