# lib/api-zod/src/generated/types/possibleCause.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PawSmart API — AI Pet Care + E-Commerce
 * OpenAPI spec version: 0.1.0
 */
import type { PossibleCauseSeverity } from "./possibleCauseSeverity";

export interface PossibleCause {
  name: string;
  severity: PossibleCauseSeverity;
  description: string;
}