# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PawSmart API — AI Pet Care + E-Commerce
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  brand: string;
  /** food | treats | accessories | grooming | toys */
  category: string;
  /** dog | cat | small_pet */
  petType: string;
  price: number;
  stock: number;
  description: string;
  image: string;
  /** e.g. 1kg, 5kg, 250g */
  weight: string;
  expiryDate: string;
  popular: boolean;
  breed?: string | null;
  /** puppy | adult | senior | kitten | all */
  ageGroup?: string | null;
  healthCondition?: string | null;
  createdAt: string;
}