# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PawSmart API — AI Pet Care + E-Commerce
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  name: string;
  brand: string;
  category: string;
  petType: string;
  price: number;
  stock: number;
  description: string;
  image: string;
  weight: string;
  expiryDate: string;
  popular?: boolean;
  breed?: string | null;
  ageGroup?: string | null;
  healthCondition?: string | null;
}